+++ test/fixtures/violations/separator_indentation.xs
@@ -1,0 +1,16 @@
+function "example" {
+  input {
+  }
+  
+  stack {
+    var $ok {
+      value = 1
+    }
+
+    var $next {
+      value = 2
+    }
+  }
+
+  response = $ok
+}